Task 3: likes and comments implemented to shorts.
save entity for liking both posts and shorts.
same with comment.
now each comment can be commented and this reply can be answered too. Fixed order of comments.

+++ social-server/src/main/java/com/mykyda/talantsocials/dto/PostDTO.java
@@ -27,7 +27,7 @@ public class PostDTO {
 
     private Timestamp createdAt;
 
-    private Integer likesAmount;
+    private Long likesAmount;
 
     public static PostDTO of(Post post) {
         return PostDTO.builder()
